# docs/swagger.yaml
basePath: /
definitions:
...
  dto.RequestComment:
    properties:
      id:
        type: string
      message:
        example: your waifuu has been claimed
        type: string
      photo_id:
        example: "1"
        type: string
      user_id:
        example: "1"
        type: string
    required:
    - message
    - photo_id
    - user_id
    type: object
  dto.RequestCommentUpdate:
    properties:
      id:
        type: string
      message:
        type: string
      user_id:
        type: string
    required:
    - message
    - user_id
    type: object
...
  dto.RequestSocialMedia:
    properties:
      name:
        type: string
      social_media_url:
        type: string
      user_id:
        type: string
    required:
    - name
    - social_media_url
    - user_id
    type: object
...
  dto.ResponseCreateComment:
    properties:
      id:
        type: string
      message:
        type: string
      photoID:
        type: string
      user_id:
        type: string
    type: object
...
  dto.ResponseCreateSocialMedia:
    properties:
      created_at:
        type: integer
      id:
        type: string
      name:
        type: string
      social_media_url:
        type: string
      updated_at:
        type: integer
      user_id:
        type: string
    type: object
  dto.ResponseGetComment:
    properties:
      id:
        type: string
      message:
        type: string
      photo:
        properties:
          caption:
            type: string
          id:
            type: string
          photo_url:
            type: string
          title:
            type: string
          user_id:
            type: string
        type: object
      photo_id:
        type: string
      user:
        properties:
          email:
            type: string
          id:
            type: string
          username:
            type: string
        type: object
      user_id:
        type: string
    type: object
...
  dto.ResponseGetSocialMedias:
    properties:
      created_at:
        type: integer
      id:
        example: "1"
        type: string
      name:
        example: kei karuizawa
        type: string
      social_media_url:
        example: https://twitter.com/keikaruizawa
        type: string
      updated_at:
        type: integer
      user:
        properties:
          id:
            type: string
          profile_image_url:
            example: https://photos.com/keikaruizawa.png
            type: string
          username:
            example: keikaruizawa
            type: string
        type: object
      user_id:
        example: "1"
        type: string
    type: object
...
  helper.ExampleErrorResponse:
    properties:
      email:
        example: cannot be empty
        type: string
      message:
        example: user bad request
        type: string
      password:
        example: cannot be empty
        type: string
    type: object
  helper.Response:
    properties:
      data: {}
      error: {}
      status:
        type: integer
    type: object
info:
  contact: {}
  description: This is API for completing hacktiv8 final project
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: GoGram API
  version: "1.0"
paths:
  /comments:
    get:
      consumes:
      - application/json
      description: Get comments
      parameters:
      - description: Bearer + user token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ResponseGetComment'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
        "401":
          description: Unauthorization
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
      summary: Get comments
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: Create a comment
      parameters:
      - description: Bearer + user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.RequestComment'
      produces:
      - application/json
      responses:
        "201":
          description: CREATED
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ResponseCreateComment'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
        "401":
          description: Unauthorization
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
      summary: Create a comment
      tags:
      - comments
  /comments/:commentID:
    delete:
      consumes:
      - application/json
      description: Delete a comment
      parameters:
      - description: Bearer + user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID of the comment
        in: path
        name: commentID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
        "401":
          description: Unauthorization
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
      summary: Delete a comment
      tags:
      - comments
    put:
      consumes:
      - application/json
      description: Update a comment
      parameters:
      - description: Bearer + user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID of the comment
        in: path
        name: commentID
        required: true
        type: integer
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.RequestCommentUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ResponseCreateComment'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
        "401":
          description: Unauthorization
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
      summary: Update a comment
      tags:
      - comments
...
  /socialmedias:
    get:
      consumes:
      - application/json
      description: Get all social medias
      parameters:
      - description: Bearer + user token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: SUCCESS
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ResponseGetSocialMedias'
                  type: array
              type: object
      summary: Get all social medias
      tags:
      - socialmedias
    post:
      consumes:
      - application/json
      description: Create social media
      parameters:
      - description: Bearer + user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.RequestSocialMedia'
      produces:
      - application/json
      responses:
        "201":
          description: CREATED
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ResponseCreateSocialMedia'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
      summary: Create new social media
      tags:
      - socialmedias
  /socialmedias/:socialmediaid:
    delete:
      consumes:
      - application/json
      description: Delete by id social media
      parameters:
      - description: Bearer + user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID of the social media
        in: path
        name: socialmediaid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: SUCCESS
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
        "404":
          description: Record not found
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
      summary: Delete by id social media
      tags:
      - socialmedias
    put:
      consumes:
      - application/json
      description: Update by id social media
      parameters:
      - description: Bearer + user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID of the social media
        in: path
        name: socialmediaid
        required: true
        type: integer
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.RequestSocialMedia'
      produces:
      - application/json
      responses:
        "200":
          description: SUCCESS
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ResponseCreateSocialMedia'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
        "404":
          description: Record not found
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
      summary: Update by id social media
      tags:
      - socialmedias
...
swagger: "2.0"